server/requirements.txt
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.13
django-rest-swagger==2.2.0
djangorestframework==3.9.4
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
openapi-codec==1.3.2
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
requests==2.24.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
uWSGI==2.0.18
zipp==3.1.0